libsrc/libbioimg/core_lib/xtypes.h: use C++11 fixed integer size types on Unix derivates

+++ libsrc/libbioimg/core_lib/xtypes.h
@@ -19,6 +19,7 @@
 #include <cmath>
 #include <vector>
 #include <limits>
+#include <cstdint>
 
 namespace bim {
 
@@ -55,10 +56,10 @@ typedef __int64 int64;
 typedef unsigned __int32 uint32;
 typedef unsigned __int64 uint64;
 #else
-typedef int int32;
-typedef long long int64;
-typedef unsigned int uint32;
-typedef unsigned long long uint64;
+typedef int32_t int32;
+typedef int64_t int64;
+typedef uint32_t uint32;
+typedef uint64_t uint64;
 #endif
 
 typedef	float float32;
